+++ lib/mix/tasks/ccxt_extract.update.ex
@@ -134,6 +134,7 @@ defmodule Mix.Tasks.CcxtExtract.Update do
     ccxt_extract.exchanges
     ccxt_extract.describe
     ccxt_extract.load_markets
+    ccxt_extract.url_templates
   )
 
   defp run_quickbeam_extractors do